voice-pipeline/src/requirements.txt
google-cloud-storage
google-cloud-bigquery
langchain
langchain-groq
requests
pydantic
dateparser